+++ backend/server.py
@@ -62,6 +62,7 @@ from routers.verdict_stage2 import router as verdict_stage2_router
 from routers.activity import router as activity_router
 from routers.incidents import router as incidents_router
 from routers.autonomous_investigator import router as autonomous_investigator_router
+from routers.autonomous_investigator import _registry_router as investigator_registry_router
 from routers.xdr_dashboard import router as xdr_dashboard_router
 from routers.xdr_mss import router as xdr_mss_router
 from routers.xdr_queue_ops import router as xdr_queue_ops_router
@@ -199,6 +200,7 @@ api.include_router(verdict_stage2_router)
 api.include_router(activity_router)
 api.include_router(incidents_router)
 api.include_router(autonomous_investigator_router)
+api.include_router(investigator_registry_router)
 api.include_router(xdr_dashboard_router)
 api.include_router(xdr_mss_router)
 api.include_router(xdr_queue_ops_router)